update test files partition
